reorganize the headers to minimize the dependencies
Until now, we had a "master" header that included the whole world: kadsim.h
Any changes to this file trigger a recompilation of all the files.

That's not a big deal for now (because the project is small) but it's
not clean.

This commit creates new header files, use forward declarations and
reorganize the includes in order to have each file only depends on what
it really needs.

That way, we will be able to really take advantage of incremental
compilation.