cleanup BitMap implementation
First of all, we rename the check method: the comment says "Check that
all bits are taken" so you expect the method to returns true when all
the bits are used (i.e the BitMap is exhausted).
Rename the method to reflects that.

Then, we mark as "const" the operations that don't modify the object
(namely `is_exhausted` and `get_bit`).

And finally, we use uint32_t instead of int (fixed-size integers should
be preferred when possible).