start an exceptions hierarchy
QuadIron should have its own exceptions hierarchy.

Also, this allows us to sidestep the false positive for `cert-err60-cpp` on the CI¹…

clang 5 on Ubuntu 14.04 uses libstdc++ 4.8 which seems to be a bit buggy
regarding `std::runtime_error` (it's not nothrow copy constructible).
This seems to be fixed since libstdc++ 5.1 (see https://stackoverflow.com/q/35039927).

Note that for now we use ugly names (with prefix) for our exceptions,
but this will change soon (when we introduce a namespace).