add a root namespace, kad, for QuadIron
It's a good (and useful!) practice to put our code in a namespace.

First benefit: we can get rid of the manual prefix "Kad" of our classes.

We will starts with a flat root namespace (kad) but as the code evolves
(split Node/Simulator) we can add sub-namespace (kad::sim for the
simulator, kad::net for the RCP communication, …).