use run-clang-tidy
`run-clang-tidy.py` is a Python script packaged with clang-tidy.

The advantages over running clang-tidy by ourself are:
- automatically run the linter on every files of the project (using the
compilation database)
- run several instance of clang-tidy in parallel (on my machine, the
time to lint the project is halved).

The drawback is that I didn't find an easy way to tell him to ignore
some files (namely our tests that use GoogleTest's macros) so we have to
manually disable the lint on the tests for now.

I think the advantages outweights the drawbacks, so let's move to
`run-clang-tidy.py`.