use const and const reference when possible
Sometimes we use pointer when what we want cannot be NULL and we are
only reading it: in these cases, a const reference (for function
parameters) or a const pointer on a const value (for object members)
makes more sense (and the compiler can enforce our implicit policy).

Note that we have to declare GethClient as mutable because even though
it is-conceptually const (we don't modify it) the methods of the client
are not const so without mutable we couldn't call them.

This is an intermediate step, in the long run we should get rid of naked
pointers anyway.