add license preamble in every file
It is good to do that for several reasons, amongst which: if someone
reuse our code or a part of it, it will be clear that our code is under
our license and not under whatever license the guy's project uses.

As the The 3-Clause BSD License is quite short, the preamble is in fact
the whole license.