introduce NodeAdress
This class contains the information required to communicate with a node.
It will be used by the Node implementation and the DHT routing
algorithm.

For now, we only have "fake" node that are simulated so IP and port are
not used (and thus not implemented, a placeholder is used instead).