uint160: implement relational operators
If we want our UInt160 to behave like builtin integers, it should
defines a total order. This is done by defining the relational operators
for our type.

Note that since the limbs are stored in "big endian" we can delegate all
the work to operators of the underlying std::array.