uint160: implement multiplication
The algorithm used is a modified version of the Algorithm M from The Art
of Computer Programming, vol. 2 by Donal Knuth.

In our case, the multiplications is between two n-places numbers and the
result is a n-place number (not an n+m numbers like in the more generic
cases).

Note that, unlike addition and subtraction, the base implementation is
not the in-place one (because you can't really implement it in place,
you have to compute the result before copying it back into *this).