uint160: implement division
The implementation is based on the long division algorithm, in its
binary version.

    if D = 0 then
      error(DivisionByZeroException)
    end
    Q := 0
    R := 0
    for i := n − 1 .. 0 do
      R := R << 1
      R(0) := N(i)
      if R ≥ D then
        R := R − D
        Q(i) := 1
      end
    end

See: https://en.wikipedia.org/wiki/Division_algorithm#Integer_division_(unsigned)_with_remainder

That's not the best/fastest algorithm, but probably the simplest.
Moreover, we don't use a lot of division in Quadiron (for now, it is
used only once in fact) so it should be good enough for now.

We can still improve it later, if need be.

Note that since we don't have methods to test/set individual bit, we
rely on bitwise operation on the UInt160 to do the job.
This probably could be improved.