use UInt160 instead of CBigNum
We don't really need integer with arbitrary precision. 160-bit unsigned
integer are enough (and if we need more, we can rework UInt160 to be
more generic, the code is simple enough).

Since the code is simpler (and we have the ownership of it) it will be
easier to maintain/ensure its correctness.

Another nice perk of a simpler code/tailored to our use is that we have
performance improvement: I measured the time it takes to initialize the
simulator (time between running quadiron with the default parameters and
having access to the shell):
- with CBignum: ~67 seconds
- with UInt160: ~8 seconds

Not bad :)

Bonus point: this allows us to get rid of OpenSSL as a dependency.