fix NTTEC URL
The NTTEC library was moved to Scality.

The one hosted by vrancurel is no longer updated.

This went unnoticed for a while, but since I updated the version of
libstdc++ to 4.9 (to support C++14) the compilation is broken (because
old NTTEC depends on GMP, and the GMP shipped on Ubuntu Trusty doesn't
work with libstdc++ > 4.8 because of [1]).

[1]: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=56019