build a docker image for the CI
Since we have to migrate to CircleCI v2, we are now allowed to use
custom docker image.
That's great as this will allow us to use a more recent version of
Ubuntu (for example, the last LTS) and thus we won't have to use several
PPAs to get up-to-date compilers (also, we can bump the version of
libjson-rpc-cpp from 0.6 to 1.0).