improve log readability
The location (using the formatter %loc) is less informative than the
function name, and can be a bit unreadable (since it use absolute path
for the file name, log messages can be very lengthy).
EasyLogging provides a %func formatter, but it is very precise and that
hurts the readability (it shows the whole function signature, not just
its name), so instead of using %func we directly use __func__ in our
logging macros.